requirements.txt
beautifulsoup4
requests
python-dotenv
moviepy
notebook
face-recognition
selenium
keras
tensorflow